cloud_ttt_e2e/requirements.txt
 torch==2.3.1+cu121
 triton==2.3.1
 einops
 tqdm
 pyyaml
 numpy
